feat: `bench check-for-release` for sanity checks
A common source of bugs is trivial things that we can check before
releasing, this applies to not only Frappe but many third party apps.
This command acts as entry-point for ALL such checks.

First cut version, no guarantees for now.
