+++ frappe/tests/test_patches.py
@@ -137,41 +137,3 @@ class TestPatchReader(FrappeTestCase):
 	def test_ignore_comments(self, _file):
 		all, pre, post = self.get_patches()
 		self.assertEqual(pre, ["app.module.patch1", "app.module.patch3"])
-
-	def test_verify_patch_txt(self):
-		"""Make sure all patches/**.py files are part of patches.txt"""
-		check_patch_files("frappe")
-
-
-# Do not remove/rename this function, other apps depend on it to test their patches
-def check_patch_files(app):
-	"""Make sure all patches/**.py files are part of patches.txt"""
-
-	patch_dir = Path(frappe.get_app_path(app)) / "patches"
-
-	app_patches = [p.split(maxsplit=1)[0] for p in patch_handler.get_patches_from_app(app)]
-
-	missing_patches = []
-
-	for file in patch_dir.glob("**/*.py"):
-		module = _get_dotted_path(file, app)
-		try:
-			patch_module = frappe.get_module(module)
-			if hasattr(patch_module, "execute"):
-				if module not in app_patches:
-					missing_patches.append(module)
-		except Exception:
-			# patch so bad it doesn't even import :shrug:
-			missing_patches.append(module)
-
-	if missing_patches:
-		raise Exception(f"Patches missing in patch.txt: \n" + "\n".join(missing_patches))
-
-
-def _get_dotted_path(file: Path, app) -> str:
-	app_path = Path(get_bench_path()) / "apps" / app
-
-	*path, filename = file.relative_to(app_path).parts
-	base_filename = Path(filename).stem
-
-	return ".".join(path + [base_filename])
